lab1/bot/requirements.txt
# Основные зависимости для Telegram-бота
python-telegram-bot[job-queue]==21.0.1
python-dotenv==1.0.0
pytz==2023.3

# Дополнительные зависимости для разработки (опционально)
# pytest==7.4.3
# black==23.9.1
# flake8==6.1.0
